Limit the maximum size of a pooled log msg
Free log messages are pooled in a channel. When logging calls are made, the
message buffers are grown to accomodate the call and, after the message
is sent to syslog, the message is re-pooled with the new size.

In cases with absurdly long logging messages, the logging pool can grow a lot,
to the point of exhausting the processes memory.

To prevent this, I added a limit to the maximum size of a pooled message.